WIP: dup returned fd
This will be passed in to driver, who expects to take ownership of the
fd.. so don't hand out our internal fd without dup()'ing it.

TODO: double check this.. EGL_EXT_image_dma_buf_import says driver does
not take ownership of the fd, and I think platform_android is using
almost the same code-paths.. so possibly the dup() should be in the
driver.  (but dri3 doesn't appear to explode.. root question is whether
screen->resource_from_handle() takes ownership of the fd or not, so not
sure if there is already a dup() somewhere in the dma_buf_import path.)